Move RpcSignatureStatus into its own module
And fixup some imports from previous commits.
